Move augur and SeqKit-related functions to shared modules
To be used by other modules.
