+++ augur/io/sequences.py
@@ -2,10 +2,17 @@ import Bio.SeqIO
 import os
 
 from augur.errors import AugurError
+from augur.utils import augur
 from importlib.metadata import version as installed_version
 from packaging.version import Version
-from typing import Iterator, Iterable, Union
+from shlex import quote as shquote
+from shutil import which
+from tempfile import NamedTemporaryFile
+from textwrap import dedent
+from typing import IO, Iterator, Iterable, Union
 from .file import open_file
+from .print import _n, indented_list
+from .shell_command_runner import run_shell_command
 
 
 def get_biopython_format(augur_format: str) -> str:
@@ -439,3 +446,98 @@ def _read_genbank(reference, feature_names):
         print(f"WARNING: {features_skipped} CDS features skipped as they didn't have a locus_tag or gene qualifier.")
 
     return features
+
+
+def validate_sequence_file(file: str):
+    """Ensure all sequence identifiers in the file are unique."""
+    with NamedTemporaryFile("w+", encoding="utf-8") as dup_num_file:
+        command = f"""
+            {augur()} read-file {shquote(file)} |
+            {seqkit()} rmdup --dup-num-file {shquote(dup_num_file.name)} > /dev/null
+        """
+
+        try:
+            run_shell_command(command, raise_errors=True)
+        except Exception as error:
+            raise AugurError(f"Validation failed for {file!r}. See error above.") from error
+
+        dup_num_file.seek(0)
+
+        if duplicates := list(sorted(parse_seqkit_dup_num_file(dup_num_file))):
+            raise AugurError(dedent(f"""\
+                Sequence ids must be unique.
+
+                The following {_n("id was", "ids were", len(duplicates))} were duplicated in {file!r}:
+
+                  {indented_list(duplicates, '                ' + '  ')}
+            """))
+
+
+def seqkit():
+    """
+    Internal helper for invoking SeqKit.
+
+    Unlike ``augur.merge.sqlite3()``, this function is not a wrapper around
+    subprocess.run. It is meant to be called without any parameters and only
+    returns the location of the executable. This is due to differences in the
+    way the two programs are invoked.
+    """
+    seqkit = os.environ.get("SEQKIT", which("seqkit"))
+    if not seqkit:
+        raise AugurError(dedent(f"""\
+            Unable to find the program `seqkit`.  Is it installed?
+            In order to use `augur merge`, SeqKit must be installed
+            separately.  It is typically provided by a Nextstrain runtime.
+            """))
+    return shquote(seqkit)
+
+
+def parse_seqkit_dup_num_file(handle: IO[str]):
+    """Extract duplicated ids from the output of seqkit rmdup --dup-num-file.
+
+    There is one line per duplicated id, in the format of a number n followed by
+    tab character followed by the id duplicated n times, separated by ", ".
+
+    Example FASTA headers
+
+        > id, 1
+        > id, 2
+        > id, 3
+        > id1
+        > id1
+
+    will produce a --dup-num-file output file
+
+        3\tid,, id,, id,
+        2\tid1, id1
+
+    which parsed by this function will yield
+
+        id,
+        id1
+
+    Examples
+    ========
+
+    Two duplicate ids.
+
+    >>> from io import StringIO
+    >>> contents = "3\\tid,, id,, id,\\n2\\tid1, id1\\n"
+    >>> list(parse_seqkit_dup_num_file(StringIO(contents)))
+    ['id,', 'id1']
+
+    One duplicate id.
+
+    >>> contents = "4\\tid, id, id, id,\\n"
+    >>> list(parse_seqkit_dup_num_file(StringIO(contents)))
+    ['id,']
+
+    An empty file returns an empty list.
+
+    >>> list(parse_seqkit_dup_num_file(StringIO("")))
+    []
+    """
+    for line in handle:
+        line = line.rstrip('\n')
+        if ", " in line:
+            yield line[line.rfind(", ")+2:]
